25-01-2021
Dany went to the kings landing supermarket and bought some groceries. As
a vendor you have to ask the numbers of items she bought and store the
itemid, itemname, cost of the items and store them in the ‘sales’ table
of database and also execute an SQL query to find the total cost and
list out the item she bought.