25-01-2021
A workshop based on web development is being conducted in the university. Now James wants to know how many members of his class are interested in it and their details.

1.He created a new table namely Student in the Oracle Database with columns: Student_ID, Student_Name, Email, Date_of_Birth.
He stored some records into the Student table.He needs to collect ID numbers and contact numbers of the interested students in a new table called"Workshop".

2.He wants to retrieve name and email address from the student table and update the Workshop table by creating 2 new columns and inserting the respective details.

3.If a student, suddenly wants to drop from attending the workshop, James is supposed to delete the student's data from workshop table.

Write the Java programs for the above situations using JDBC API.